Changed all toast heights to display above the now playing bar | Changed default album art colour to match other icons

+++ app/src/main/java/gg/joshbra/tagg/Fragments/BottomTaggSelectionFragment.java
@@ -47,7 +47,7 @@ public class BottomTaggSelectionFragment extends BottomSheetDialogFragment {
             @Override
             public void onClick(View view) {
                 final EditText newTaggEditText = new EditText(view.getContext());
-                newTaggEditText.setTextColor(getResources().getColor(R.color.colorTextSecondary));
+                newTaggEditText.setTextColor((int) getResources().getColor(R.color.colorTextSecondary));
 
                 newTaggEditText.requestFocus();
